estuve trabajando en el tema de las creaciones y los productos

+++ src/main/java/com/example/PizzUMBurgUM/controladores/ControladorProducto.java
@@ -24,7 +24,7 @@ public class ControladorProducto {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Producto> obtenerPorId(@PathVariable("id") Integer id) {
+    public ResponseEntity<Producto> obtenerPorId(@PathVariable("id") Long id) {
         return ResponseEntity.ok(productoServicio.obtenerProductoPorId(id));
     }
 }
