+++ src/main/java/com/example/PizzUMBurgUM/repositorios/CreacionRepositorio.java
@@ -3,5 +3,5 @@ package com.example.PizzUMBurgUM.repositorios;
 import com.example.PizzUMBurgUM.entidades.Creacion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CreacionRepositorio extends JpaRepository<Creacion, Integer> {
+public interface CreacionRepositorio extends JpaRepository<Creacion, Long> {
 }
