+++ src/main/java/com/example/PizzUMBurgUM/repositorios/ProductoRepositorio.java
@@ -3,5 +3,5 @@ package com.example.PizzUMBurgUM.repositorios;
 import com.example.PizzUMBurgUM.entidades.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductoRepositorio extends JpaRepository<Producto, Integer> {
+public interface ProductoRepositorio extends JpaRepository<Producto, Long> {
 }
